+++ src/components/Navbar/styles.module.scss
@@ -46,16 +46,7 @@
       }
     }
     .dropdown {
-      cursor: pointer;
-      margin-right: 2.5rem;
-      font-weight: 500;
-      font-size: 16px;
-      color: $primary_color;
-      position: relative;
-
-      &:hover {
-        color: $accent_color;
-      }
+      @include dropdown;
     }
 
     .dropdown_content {
@@ -83,32 +74,6 @@
         }
       }
     }
-    .dropdown_lang {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      position: absolute;
-      max-width: 100%;
-      background-color: white;
-      border: 3px solid $accent_gradient;
-      border-radius: 8px;
-      box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-      padding: 0.4rem;
-      z-index: 1;
-      margin-top: 1rem;
-
-      .link {
-        color: $accent_medium;
-        text-decoration: none;
-        margin: 0.5rem;
-        transition: color 0.3s ease;
-      }
-
-      .link:hover {
-        color: $accent_color;
-      }
-    }
-
     .nav_about {
       display: flex;
       justify-content: space-evenly;
